+++ apps/frontend/src/pages/landing/libs/components/popular-section/popular-section.tsx
@@ -1,5 +1,3 @@
-import { RouteCard } from "~/libs/components/components.js";
-
 import styles from "./styles.module.css";
 
 const PopularSection = (): React.JSX.Element => {
@@ -7,14 +5,7 @@ const PopularSection = (): React.JSX.Element => {
 		<section className={styles["section"]}>
 			<div className={styles["container"]}>
 				<h2 className={styles["title"]}>Popular routes</h2>
-				<ul className={styles["cards"]}>
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-					<RouteCard imageUrl={null} name="Kralovska obora" />
-				</ul>
+				<ul className={styles["cards"]} />
 			</div>
 		</section>
 	);