+++ packages/shared/src/modules/files/libs/types/file-upload-response-dto.type.ts
@@ -2,6 +2,7 @@ import { type FileMimeType } from "./file-mime-type.type.js";
 
 type FileUploadResponseDto = {
 	contentType: FileMimeType;
+	createdAt: string;
 	id: number;
 	url: string;
 };