+++ packages/shared/src/modules/routes/routes.ts
@@ -3,6 +3,7 @@ export {
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,
 	type RouteCreateRequestDto,
+	type RouteCreateResponseDto,
 	type RouteFindAllOptions,
 	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,